Align test with model used for E2E validations in Azure Quantum

+++ src/quantum/azext_quantum/tests/latest/test_quantum_workspace.py
@@ -9,7 +9,7 @@ import unittest
 from azure_devtools.scenario_tests import AllowLargeResponse, live_only
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer)
 
-from .utils import get_test_resource_group, get_test_workspace, get_test_workspace_location, get_test_workspace_storage, get_test_workspace_random_name
+from .utils import get_test_resource_group, get_test_workspace, get_test_workspace_location, get_test_workspace_storage, get_test_workspace_random_name, get_test_capabilities, get_test_workspace_provider_sku_list, all_providers_are_in_capabilities
 from ..._version_check_helper import check_version
 from datetime import datetime
 from ...__init__ import CLI_REPORTED_VERSION 
@@ -55,18 +55,22 @@ class QuantumWorkspacesScenarioTest(ScenarioTest):
         test_resource_group = get_test_resource_group()
         test_workspace_temp = get_test_workspace_random_name()
         test_storage_account = get_test_workspace_storage()
+        test_provider_sku_list = get_test_workspace_provider_sku_list()
 
-        # create
-        self.cmd(f'az quantum workspace create -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account} -r "Microsoft/Basic" -o json --skip-role-assignment', checks=[
-           self.check("name", test_workspace_temp),
-           self.check("provisioningState", "Accepted")  # Status is accepted since we're not linking the storage account.
-        ])
+        if all_providers_are_in_capabilities(test_provider_sku_list, get_test_capabilities()):
+            # create
+            self.cmd(f'az quantum workspace create -g {test_resource_group} -w {test_workspace_temp} -l {test_location} -a {test_storage_account} -r {test_provider_sku_list} -o json --skip-role-assignment', checks=[
+            self.check("name", test_workspace_temp),
+            self.check("provisioningState", "Accepted")  # Status is accepted since we're not linking the storage account.
+            ])
 
-        # delete
-        self.cmd(f'az quantum workspace delete -g {test_resource_group} -w {test_workspace_temp} -o json', checks=[
-           self.check("name", test_workspace_temp),
-           self.check("provisioningState", "Deleting")
-        ])
+            # delete
+            self.cmd(f'az quantum workspace delete -g {test_resource_group} -w {test_workspace_temp} -o json', checks=[
+            self.check("name", test_workspace_temp),
+            self.check("provisioningState", "Deleting")
+            ])
+        else:
+            self.skipTest(f"Skipping test_workspace_create_destroy: One or more providers in '{test_provider_sku_list}' not found in AZURE_QUANTUM_CAPABILITIES")
 
     @live_only()
     def test_version_check(self):
@@ -86,4 +90,3 @@ class QuantumWorkspacesScenarioTest(ScenarioTest):
 
         message = check_version(test_config, test_none_version, test_today)
         assert message is None  # Note: list_versions("quantum") fails during these tests
-
